program now running

+++ banksStatus.c
@@ -1,4 +1,4 @@
 #define numOfBnaks 50
 
 
-int banks_status[numOfBnaks][2] = {0};
+double banks_status[numOfBnaks][2] = {0};
